Update minify.php - Do not remove space, before (
Line 29 and 30. 

Found this out today...you need the space, before ( , if you are using media queries, otherwise the media query will not initialize. 

The previous would take: 

@media screen and (max-width: XXXpx){
}
and turn it into 
@media screen and(max-width: XXXpx){}

the and(  causes the media query to fail. 

Also thank you for this file...found it last week and it has been an amazing help, so far. 

(in my proposed change, I simply removed |\( from line 30. )